projects/marketing_agency/archive/sample1_naver-diagnosis_legacy/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
playwright==1.44.0
python-pptx==0.6.23
httpx==0.27.0
jinja2==3.1.4
python-multipart==0.0.9
openpyxl==3.1.2
anthropic>=0.40.0
google-generativeai>=0.8.0
